requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
autopep8==2.0.1
backcall==0.2.0
boto3==1.26.94
botocore==1.29.94
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
cryptography==39.0.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dj-database-url==1.2.0
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.53.1
django-braces==1.15.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==22.1
django-import-export==3.1.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-leaflet==0.28.3
django-oauth-toolkit==2.2.0
django-redis==5.2.0
django-storages==1.13.2
django-taggit==3.1.0
django-tinymce==3.5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
executing==1.2.0
Faker==18.3.2
geographiclib==2.0
geopy==2.3.0
graphviz==0.20.1
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
jwcrypto==1.4.2
Markdown==3.4.1
MarkupPy==1.14
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
redis==4.5.1
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tablib==3.4.0
threadpoolctl==3.1.0
timedelta==2020.12.3
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
urllib3==1.26.15
wcwidth==0.2.6
whitenoise==6.4.0
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
